+++ src/app/app.module.ts
@@ -8,22 +8,29 @@ import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { MaterialModule } from './material';
 import { DashboardModule } from 'src/app/modules/dashboard/dashboard.module';
+import { HttpUtilsModule } from 'src/app/modules/http-utils/http-utils.module';
 import { ToastrModule } from 'ngx-toastr';
+import { HttpModule } from '@angular/http';
+
+
  @NgModule({
   declarations: [
     AppComponent
   ],
   imports: [
     BrowserModule,
+    HttpUtilsModule,
+    HttpModule,
     HttpClientModule,
     AppRoutingModule,
     BrowserAnimationsModule,
     MaterialModule,
     DashboardModule,
     RouterModule,
-    ToastrModule.forRoot(),
+    ToastrModule.forRoot()
    ],
-  providers: [],
-  bootstrap: [AppComponent]
+  providers: [ HttpUtilsModule],
+  bootstrap: [AppComponent],
+   entryComponents: []
 })
 export class AppModule { }